deep-model/requirements.txt
# pip3 install -U -r requirements.txt
# conda install numpy opencv matplotlib tqdm pillow && conda install pytorch torchvision -c pytorch
numpy
mxnet
gluoncv
matplotlib
grpcio
grpcio-tools
